companymgmt/requirements.txt
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
distlib==0.3.4
Django==4.0.2
django-filter==21.1
django-stubs==1.10.1
django-stubs-ext==0.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
djangorestframework-stubs==1.5.0
dparse==0.5.1
filelock==3.6.0
flake8==4.0.1
identify==2.4.12
idna==3.3
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.941
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.3
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
pre-commit==2.18.1
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
safety==1.10.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
types-Markdown==3.3.12
types-pytz==2021.3.6
types-PyYAML==6.0.5
types-requests==2.27.16
types-urllib3==1.26.11
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
